backend/requirements.txt
fastapi[standard]>=0.115.6,<0.116.0
pydantic[email]>=2.9.2,<3.0.0
python-dotenv>=1.0.1,<1.1.0
# httpx>=0.28.0,<0.29.0
sqlalchemy>=2.0.36,<2.1.0
PyMySQL>=1.1.1,<1.2.0
alembic>=1.14.0,<1.15.0
